+++ backend/buddies/admin.py
@@ -1,0 +1,5 @@
+import django.contrib.admin
+
+import buddies.models
+
+django.contrib.admin.site.register(buddies.models.Buddy)