+++ node/packages/core/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./utils.ts";
 export * from "./types";
 export * from "./canvas";
-export * from "./constants";
 export * from "./skills/createTasks";
 export * from "./skills/executeTasks";
 export * from "./skills/utils";