+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - Support Azure Blob Storage as object storage backend.
+- Available LLMs can now be configured in the backend.
 
 ## [25.5.0] - 2025-05-05